+++ R/Discrete-methods.R
@@ -39,7 +39,7 @@ setMethod(
             cat("\nLog-likelihood = ", x@Fit$logLik, if(method == 'MHRM')
                 paste(', SE =', round(x@Fit$SElogLik,3)), "\n",sep='')
             cat('Estimated parameters:', extract.mirt(x, 'nestpars'), '\n')
-            cat("AIC = ", x@Fit$AIC, "; AICc = ", x@Fit$AICc, "\n", sep='')
+            cat("AIC = ", x@Fit$AIC, "\n", sep='')
             cat("BIC = ", x@Fit$BIC, "; SABIC = ", x@Fit$SABIC, "\n", sep='')
             if(!is.nan(x@Fit$p)){
                 cat("G2 (", x@Fit$df,") = ", round(x@Fit$G2,2), ", p = ", round(x@Fit$p,4), sep='')